auth-service/requirements.txt
fastapi
uvicorn
pyjwt[crypto]
sqlmodel
pydantic-settings
asyncpg
passlib[argon2]
pytest
httpx
aiosqlite
pytest-asyncio
greenlet